+++ tests/test_cli.py
@@ -8,29 +8,24 @@ from click.testing import CliRunner
 from chaostoolkit.cli import cli
 
 
-def test_path_to_experiment_description_is_mandatory():
+def test_source_experiment_is_mandatory():
     runner = CliRunner()
     result = runner.invoke(cli, ['run'])
     assert result.exit_code == 2
     assert result.exception
-    assert 'Error: Missing argument "path".' in result.output
+    assert 'Error: Missing argument "source".' in result.output
 
 
-def test_path_to_experiment_description_is_mandatory():
+def test_source_path_must_exist(log_file):
     runner = CliRunner()
-    result = runner.invoke(cli, ['run', 'invalid.jsn'])
-    assert result.exit_code == 2
+    result = runner.invoke(cli, [
+        '--log-file', log_file.name, 'run', 'invalid.jsn'])
+    assert result.exit_code == 1
     assert result.exception
-    assert 'Error: Invalid value for "path": Path '\
-           '"invalid.jsn" does not exist.' in result.output
 
-
-def test_path_to_experiment_description_is_mandatory():
-    runner = CliRunner()
-    result = runner.invoke(cli, ['run', 'invalid.jsn'])
-    assert result.exit_code == 2
-    assert result.exception
-    assert 'Error: Invalid value for "path": Path "invalid.jsn" does not exist.' in result.output
+    log_file.seek(0)
+    log = log_file.read().decode('utf-8')
+    assert 'Path "invalid.jsn" does not exist.' in log
 
 
 def test_default_settings_file(log_file):